Day07 Assinment done

+++ Day07/Q25.cpp
@@ -1,6 +1,6 @@
 #include<iostream>
 using namespace std;
-int factorial(int n)
+long long factorial(int n)
 {
 if(n==1 || n==0)
 return 1;
